requirements.txt
dash==2.0.0
dash_bootstrap_components==1.0.2
dash-renderer==1.9.1
gunicorn==20.1.0
gcsfs==2021.11.1
zarr==2.10.3
numpy==1.21.5
xarray[zarr]==0.20.2
google-cloud==0.34.0
datashader==0.11.1
plotly==5.5.0
colorcet==3.0.0
# pyproj==3.2.1
flask_caching==1.10.1
google-cloud-profiler
# rioxarray==0.9.1
rasterio==1.2.10
